Fix stale exit prices, size entries notionally, recalibrate the checkpoint
Three fixes from the investigation into the open readiness issues.

1. Position management ran on yesterday's prices. The 9:30 scan reads yfinance
   daily bars before today's bar exists, so Close.iloc[-1] was still the prior
   session's close: 49 of 87 exits since April match the prior close within 0.1%.
   Stop checks, trailing updates and exits were all a session behind, and with
   orders now reaching Alpaca the exit intended_price would have measured
   staleness rather than execution. prices.py gains get_latest_price() (Alpaca
   last trade, then yfinance intraday, then the daily bar), get_prior_close()
   and get_today_open(), which resolve the current-session ambiguity explicitly.
   The entry signal now uses the opening print vs the prior close — the same
   overnight gap the backtest measures — while the fill uses the live quote.

   Repricing exits at the real open costs -0.18%/trade; entries recorded ~2
   minutes after the open sit +0.46% above it, about a quarter of the backtest's
   +2.46%/trade expectancy. Both are documented in ROADMAP.

2. Entries are sized notionally where the broker allows it. 94 of the 98 tickers
   traded since April are fractionable at Alpaca, so a $500 slot buys $500 of an
   $1,800 stock instead of one oversized share; whole-share symbols still round
   down and warn. Position and OrderResult quantities are now floats.

3. Checkpoint criteria recalibrated for statistical power. At n=60 a t-stat of 2
   needs +3.3%/trade and at n=88 it needs +2.74% — both above what the backtest
   itself produced, so the old rule tested the sample size rather than the
   strategy (the backtest's own edge scores t=1.79 at n=88). Decision now waits
   for n>=150, adds a PILOT tier at half capital, and kills on evidence — a
   negative mean or a 95% CI that excludes a 1%/trade edge — instead of on a weak
   t-stat. Current verdict: EXTEND, 62 trades short of the sample floor.

Also fixes a bug in the risk controls from the previous commit: mark_to_market()
ran before the epoch existed on the first evaluation and marked the entire legacy
book, inflating peak equity to $5,676 and tripping the drawdown breaker on the
next run. It now returns zero with no epoch, and the persisted anchors were
re-baselined to the $5,000 capital base. 125 tests pass.

+++ src/state.py
@@ -28,7 +28,7 @@ class Position:
     current_stop: float
     entry_date: str   # 'YYYY-MM-DD'
     day_count: int    # trading days held so far
-    quantity: int = 0 # shares held (set at buy time)
+    quantity: float = 0.0  # shares held (set at buy time; fractional where supported)
     mode: str = "sim" # venue the entry executed in: 'sim' | 'paper' | 'live'.
                       # Exits are routed back to the same venue, so a simulated
                       # position is never sold into a broker that does not hold it.
